Extracted out the implementation into a separate file
